Code smells fixes

+++ src/main/java/smarthome/model/Fridge.java
@@ -49,7 +49,7 @@ public class Fridge implements DeviceSpecs {
 
     @Override
     public double getEnergyConsumption() {
-        return mAnnualEnergyConsumption / 365;
+        return (double) mAnnualEnergyConsumption / 365;
     }
 
     public void setFreezerCapacity(int freezerCapacity) {
